Updated lava class; separated calcs & draws

+++ FrameTime.pde
@@ -1,10 +1,18 @@
 class FrameTime {
 
-    double currentTime;
-    double previousTime;
+    float currentTime;
+    float previousTime;
     
-    double deltaTime() {
-        double deltaTime;
+//    double deltaTime() {
+//        double deltaTime;
+//        currentTime = millis()/1000f;
+//        deltaTime = currentTime - previousTime;
+//        previousTime = currentTime;
+//        return deltaTime;
+//    }
+
+    float deltaTime() {
+        float deltaTime;
         currentTime = millis()/1000f;
         deltaTime = currentTime - previousTime;
         previousTime = currentTime;
